+++ src/app.js
@@ -1,1 +1,0 @@
-export { createApp, startServer } from "./core/app.js";