+++ src/repositories/clients.repo.js
@@ -1,1 +1,0 @@
-export { clientsRepository } from "../modules/client-projects/clients.repo.js";