+++ src/repositories/projects.repo.js
@@ -1,1 +1,0 @@
-export { projectsRepository } from "../modules/client-projects/projects.repo.js";