+++ src/routes/clients.routes.js
@@ -1,1 +1,0 @@
-export { clientsRoutes } from "../modules/client-projects/clients.routes.js";