+++ src/routes/projects.routes.js
@@ -1,1 +1,0 @@
-export { projectsRoutes } from "../modules/client-projects/projects.routes.js";