+++ src/services/clients.service.js
@@ -1,1 +1,0 @@
-export { clientsService } from "../modules/client-projects/clients.service.js";